Only initiate tabs if there's tabs to initiate (prevent plugin conflict)
I use Download Monitor (https://www.download-monitor.com) on my site, and the Insert Download popup was having a JavaScript error since this plugin was trying to initiate .tabs(); even though there weren't any tabs present (and jQuery UI hadn't been called).

This is quick fix that checks to make sure Post Snippet tabs are present before trying to initiate jQuery UI tabs. A more permanent fix might be to resolve why this plugin's script being executed when a different plugin is opening a popup in the editor, but I'm not familiar enough with this plugin's code to dive into that. This is just a quick & harmless fix.