Add conditional code for checking if dialog and tabs functions exist
I had a site where dialog wasn't being loaded correctly due to a theme compatibility issue and this plugin tried calling a function that then didn't exist & resulted in a JavaScript error. As a precaution on this plugin's end, this makes sure the functions exist before trying to use them.

Also, I reformatted the whitespace so indenting made more sense.